feat: Make Transaction

+++ database/migrations/2024_12_05_052717_create_ticketcategories_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->double('price');
             $table->foreignId('event_id')->constrained()->onDelete('cascade');
             $table->dateTime('deadline');
-            $table->bigInteger('stock');
+            $table->integer('stock');
             $table->timestamps();
         });
     }
